+++ modules/core/src/main/scala/turbolift/internals/extensions/MiscExtensions.scala
@@ -1,4 +1,4 @@
-package turbolift.internals.extensions.orphaned
+package turbolift.internals.extensions
 import turbolift.!!
 import turbolift.effects.Each
